feat(RUN-789): migrate tool loading to ToolScanner
