+++ packages/core/src/runsight_core/yaml/discovery.py
@@ -14,11 +14,9 @@ Usage:
 
 from __future__ import annotations
 
-import ast
 import importlib.util
 import logging
 import sys
-from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, Collection, Dict, Tuple
 
@@ -34,28 +32,22 @@ if __spec__ is not None:
     __spec__.submodule_search_locations = __path__
 
 from runsight_core.yaml.discovery._base import BaseScanner, ScanIndex
+from runsight_core.yaml.discovery._tool import (
+    RESERVED_BUILTIN_TOOL_IDS,
+    ToolMeta,
+    ToolScanner,
+)
 
-
-@dataclass
-class ToolMeta:
-    """Metadata for a discovered custom tool definition file."""
-
-    tool_id: str
-    file_path: Path
-    version: str
-    type: str
-    executor: str
-    name: str
-    description: str
-    parameters: dict[str, Any]
-    code: str | None = None
-    code_file: str | None = None
-    request: dict[str, Any] | None = None
-    timeout_seconds: int | None = None
-
-
-RESERVED_BUILTIN_TOOL_IDS = frozenset({"http", "file_io", "delegate"})
 logger = logging.getLogger(__name__)
+
+__all__ = [
+    "RESERVED_BUILTIN_TOOL_IDS",
+    "SoulScanner",
+    "ToolMeta",
+    "ToolScanner",
+    "_to_snake_case",
+    "discover_custom_assets",
+]
 
 
 def _to_snake_case(name: str) -> str:
@@ -72,27 +64,6 @@ def _to_snake_case(name: str) -> str:
             result.append("_")
         result.append(char.lower())
     return "".join(result)
-
-
-def _validate_tool_main_contract(code: str) -> None:
-    """Require a ``def main(args)`` function in custom tool code."""
-    try:
-        tree = ast.parse(code)
-    except SyntaxError as exc:
-        raise ValueError(f"Tool code has a syntax error: {exc}") from exc
-
-    for node in ast.walk(tree):
-        if isinstance(node, (ast.FunctionDef, ast.AsyncFunctionDef)) and node.name == "main":
-            args = node.args.args
-            if len(args) != 1 or args[0].arg != "args":
-                raise ValueError("Tool code must define 'def main(args)'")
-            return
-
-    raise ValueError("Tool code must define 'def main(args)'")
-
-
-def _fail_tool_file(yaml_file: Path, message: str) -> ValueError:
-    return ValueError(f"{yaml_file.name}: {message}")
 
 
 def _fail_soul_file(yaml_file: Path, message: str) -> ValueError:
@@ -145,190 +116,6 @@ class SoulScanner(BaseScanner[Soul]):
         for soul_key in sorted(ignored_soul_keys & set(index.stems())):
             logger.warning("Inline soul '%s' overrides external soul file", soul_key)
         return index.without_stems(ignored_soul_keys)
-
-
-def _require_string(raw: dict[str, Any], key: str, *, yaml_file: Path) -> str:
-    value = raw.get(key)
-    if not isinstance(value, str) or not value.strip():
-        raise _fail_tool_file(yaml_file, f"missing or invalid {key!r}")
-    return value
-
-
-def _require_mapping(raw: dict[str, Any], key: str, *, yaml_file: Path) -> dict[str, Any]:
-    value = raw.get(key)
-    if not isinstance(value, dict):
-        raise _fail_tool_file(yaml_file, f"missing or invalid {key!r}")
-    return value
-
-
-def _read_tool_code_file(yaml_file: Path, code_file: str) -> str:
-    code_path = yaml_file.parent / code_file
-    if not code_path.exists():
-        raise _fail_tool_file(yaml_file, f"referenced code_file does not exist: {code_file}")
-    if not code_path.is_file():
-        raise _fail_tool_file(yaml_file, f"referenced code_file is not readable: {code_file}")
-
-    try:
-        return code_path.read_text(encoding="utf-8")
-    except OSError as exc:
-        raise _fail_tool_file(
-            yaml_file, f"referenced code_file is not readable: {code_file}"
-        ) from exc
-
-
-def _normalize_request_config(yaml_file: Path, raw_request: dict[str, Any]) -> dict[str, Any]:
-    allowed_fields = {"method", "url", "headers", "body_template", "response_path"}
-    extra_fields = sorted(set(raw_request.keys()) - allowed_fields)
-    if extra_fields:
-        joined = ", ".join(extra_fields)
-        raise _fail_tool_file(yaml_file, f"unsupported request fields: {joined}")
-
-    method = raw_request.get("method", "GET")
-    url = raw_request.get("url")
-    headers = raw_request.get("headers")
-    body_template = raw_request.get("body_template")
-    response_path = raw_request.get("response_path")
-
-    if not isinstance(method, str) or not method.strip():
-        raise _fail_tool_file(yaml_file, "missing or invalid request.method")
-    if not isinstance(url, str) or not url.strip():
-        raise _fail_tool_file(yaml_file, "missing or invalid request.url")
-    if headers is not None:
-        if not isinstance(headers, dict) or any(
-            not isinstance(key, str) or not isinstance(value, str) for key, value in headers.items()
-        ):
-            raise _fail_tool_file(yaml_file, "request.headers must be a mapping of strings")
-    if body_template is not None and not isinstance(body_template, str):
-        raise _fail_tool_file(yaml_file, "request.body_template must be a string")
-    if response_path is not None and not isinstance(response_path, str):
-        raise _fail_tool_file(yaml_file, "request.response_path must be a string")
-
-    return {
-        "method": method,
-        "url": url,
-        "headers": headers or {},
-        "body_template": body_template,
-        "response_path": response_path,
-    }
-
-
-def discover_custom_tools(base_dir: str | Path) -> Dict[str, ToolMeta]:
-    """Discover custom tool metadata files from ``custom/tools/*.yaml``."""
-    base_path = Path(base_dir)
-    tools_dir = base_path / "custom" / "tools"
-
-    if not tools_dir.exists():
-        return {}
-
-    discovered: Dict[str, ToolMeta] = {}
-
-    for yaml_file in tools_dir.glob("*.yaml"):
-        tool_id = yaml_file.stem
-        if tool_id in RESERVED_BUILTIN_TOOL_IDS:
-            collision_path = yaml_file
-            try:
-                collision_path = yaml_file.relative_to(base_path)
-            except ValueError:
-                pass
-            raise _fail_tool_file(
-                yaml_file,
-                f"reserved builtin tool id {tool_id!r} collides with custom tool metadata at "
-                f"{collision_path}",
-            )
-        if tool_id in discovered:
-            raise _fail_tool_file(yaml_file, f"duplicate custom tool id collision for {tool_id!r}")
-
-        try:
-            raw = yaml.safe_load(yaml_file.read_text(encoding="utf-8"))
-        except (OSError, yaml.YAMLError) as exc:
-            raise _fail_tool_file(yaml_file, "malformed YAML") from exc
-
-        if not isinstance(raw, dict):
-            raise _fail_tool_file(yaml_file, "invalid tool metadata")
-
-        allowed_fields = {
-            "version",
-            "type",
-            "executor",
-            "name",
-            "description",
-            "parameters",
-            "code",
-            "code_file",
-            "request",
-            "timeout_seconds",
-        }
-        extra_fields = sorted(set(raw.keys()) - allowed_fields)
-        if extra_fields:
-            joined = ", ".join(extra_fields)
-            raise _fail_tool_file(yaml_file, f"unsupported fields: {joined}")
-
-        version = _require_string(raw, "version", yaml_file=yaml_file)
-        tool_type = _require_string(raw, "type", yaml_file=yaml_file)
-        if tool_type != "custom":
-            raise _fail_tool_file(yaml_file, "type must be 'custom'")
-
-        executor = _require_string(raw, "executor", yaml_file=yaml_file)
-        name = _require_string(raw, "name", yaml_file=yaml_file)
-        description = _require_string(raw, "description", yaml_file=yaml_file)
-        parameters = _require_mapping(raw, "parameters", yaml_file=yaml_file)
-        code = raw.get("code")
-        code_file = raw.get("code_file")
-        request = raw.get("request")
-        timeout_seconds = raw.get("timeout_seconds")
-        if code is not None and not isinstance(code, str):
-            raise _fail_tool_file(yaml_file, "code must be a string")
-        if code_file is not None and not isinstance(code_file, str):
-            raise _fail_tool_file(yaml_file, "code_file must be a string")
-        if timeout_seconds is not None and (
-            not isinstance(timeout_seconds, int)
-            or isinstance(timeout_seconds, bool)
-            or timeout_seconds < 1
-        ):
-            raise _fail_tool_file(yaml_file, "timeout_seconds must be a positive integer")
-
-        normalized_request: dict[str, Any] | None = None
-        if executor == "python":
-            if request is not None or timeout_seconds is not None:
-                raise _fail_tool_file(yaml_file, "python executor cannot declare request fields")
-            if code and code_file:
-                raise _fail_tool_file(
-                    yaml_file, "python executor cannot declare both code and code_file"
-                )
-            if code_file:
-                code = _read_tool_code_file(yaml_file, code_file)
-            elif not code:
-                raise _fail_tool_file(yaml_file, "python executor requires code or code_file")
-
-            try:
-                _validate_tool_main_contract(code)
-            except ValueError as exc:
-                raise _fail_tool_file(yaml_file, str(exc)) from exc
-        elif executor == "request":
-            if code is not None or code_file is not None:
-                raise _fail_tool_file(yaml_file, "request executor cannot declare python fields")
-            if not isinstance(request, dict):
-                raise _fail_tool_file(yaml_file, "request executor requires a request mapping")
-            normalized_request = _normalize_request_config(yaml_file, request)
-        else:
-            raise _fail_tool_file(yaml_file, f"unknown executor {executor!r}")
-
-        discovered[tool_id] = ToolMeta(
-            tool_id=tool_id,
-            file_path=yaml_file,
-            version=version,
-            type=tool_type,
-            executor=executor,
-            name=name,
-            description=description,
-            parameters=parameters,
-            code=code,
-            code_file=code_file,
-            request=normalized_request,
-            timeout_seconds=timeout_seconds,
-        )
-
-    return discovered
 
 
 def _discover_blocks(blocks_dir: Path) -> Dict[str, type]:
